Add Stable Master NPC Template
Tried to use Ago's modloader support but found that it makes changes to traits even if you don't use them and this might be what is messing with some servers with white foals. Also, it makes it hard to specify a template ID if you don't want one auto-created as a bunch of 'setter' functions are private.