WIP - Fixes for server transfer
Added a bunch of logging to track down some fixes. Was skipping some steps on receiving side due to being inside a conditional accidentally. Was accidentally sending inventory item. Wasn't creating a possessions object so the creature couldn't be saved.  Wasn't telling creature it was already saved after saving it.